Add pulpcore_minor_release and pulpcore_update parameters
`pulpcore_minor_release` defines the 3.Y release that should be installed.

The first time the installer for this minor release of pulpcore is run, the
latest 3.Y.Z version of pulpcore is installed. On subsequent runs of the
installer, the user must specify `pulpcore_update` as true in order to update
pulpcore to the latest bug fix release for that minor release.

fixes: #8847
https://pulp.plan.io/issues/8847